+++ frontend/components/stops/StopsListViewFavorites/index.tsx
@@ -12,7 +12,7 @@ import styles from './styles.module.css';
 
 /* * */
 
-export default function Component() {
+export function StopsListViewFavorites() {
 	//
 
 	//
