+++ frontend/components/stops/StopsListViewSkeleton/styles.module.css
@@ -1,0 +1,13 @@
+/* * */
+/* CONTAINER */
+
+.container {
+	display: flex;
+	flex-direction: column;
+}
+
+@media screen and (max-width: 500px) {
+	.container {
+		border-bottom: 1px solid var(--color-system-border-100);
+	}
+}
